chore: bump version to 0.3.62 and update telegramId validation
- Updated version in package.json to 0.3.62.
- Modified telegramId field in CreateUserCommand, UpdateUserCommand, BulkUpdateUsersCommand, BulkAllUpdateUsersCommand, and UsersSchema to enforce integer validation.

+++ libs/contract/commands/users/bulk-all/bulk-all-update-users.command.ts
@@ -37,7 +37,7 @@ export namespace BulkAllUpdateUsersCommand {
                 .describe('Expiration date: 2025-01-17T15:38:45.065Z'),
         ),
         description: z.optional(z.string().nullable()),
-        telegramId: z.optional(z.number().nullable()),
+        telegramId: z.optional(z.number().int().nullable()),
         email: z.optional(z.string().email('Invalid email format').nullable()),
     });
