+++ libs/contract/commands/users/update-user.command.ts
@@ -49,7 +49,7 @@ export namespace UpdateUserCommand {
             .describe('Expiration date: 2025-01-17T15:38:45.065Z')
             .optional(),
         description: z.optional(z.string().nullable()),
-        telegramId: z.optional(z.number().nullable()),
+        telegramId: z.optional(z.number().int().nullable()),
         email: z.optional(z.string().email('Invalid email format').nullable()),
     });
